requirements.txt
djangorestframework==3.13.1
Pillow==9.2.0
Django==4.0.0
celery==5.2.7
django-heroku==0.3.1
whitenoise==6.0.0
gunicorn==20.1.0
Redis==4.3.4